Split tests.conftest module
